:sparkles: Add an AST display of the code too
It's optional, and hidden by default. Also, at this point, it has no
interaction with the rest of the display other than it updates when the code
is updated.

Next up is seeing if we can get the source, the bytecode and the AST to all
update each other in respect to location.