:bug: Handle argval possibly not being populated
With Python 3.14 at least one opcode (LOAD_SMALL_INT) seems to not populate
argrepr when there is an argval. This is quite different from all other
opcodes I've seen so far. I suspect this might be a bug with Python 3.14's
dis module, but that's a digging session for another time (to be clear:
dis.dis shows the value; but the Instruction from Bytecode doesn't have a
populated argrepr).

So here I change things so that I use argrepr if it is populated, otherwise
I repr argval if it isn't None. Technically this could result in a false
negative if there's a genuine None as the argval, but then I'd hope that
argrepr is populated with "None".